requirements.txt
# ─── Core RAG ───────────────────────────────────────────────
chromadb==0.5.3
sentence-transformers==3.0.1
torch==2.3.1

# ─── Groq LLM ───────────────────────────────────────────────
groq==0.8.0

# ─── Enriquecimiento liviano (sin crawl4ai) ─────────────────
requests==2.31.0
beautifulsoup4==4.12.3
lxml>=5.3.0

# ─── Web App ────────────────────────────────────────────────
flask==3.0.3
gunicorn==22.0.0

# ─── Utils ──────────────────────────────────────────────────
python-dotenv==1.0.1
Pillow==10.4.0
tqdm==4.66.4
httpx==0.27.2
